remove old FormRadioGroup
